Move Navigator interfaces in their own file
